requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jupyter-client==6.0.0
jupyter-core==4.6.3
kiwisolver==1.2.0
matplotlib==3.3.1
networkx==2.5
numpy==1.19.1
pandas==1.1.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.0
scikit-image==0.17.2
scipy==1.5.2
six==1.14.0
tifffile==2020.8.25
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.8
